Balance Update (Will Require Play-Testing)
@githubupdates THE MATH HAS BEEN DONE

Even with the prior updates, the Monk STILL fell behind basic options for other martial characters, with some cases being that it fell behind by 30-40+ average DPR (Damage per round) comparatively to other classes like the Paladin and The Fighter. As such, a huge balance re-adjustment was made to the Monk.

The Rogue was over-adjusted (which was my bad) and new scaling has been made to the Sneak Attack Dice. This will currently be in testing phases, but we will have to see how it holds up with the still decent critical hit rate.

A new feat has been added, Ki Expert, which gives the Monk a couple of extra Ki points every time it is taken, or allows other classes to take a small dip into Ki and try it out. It gives two options to use Ki for which can increase damage for a weapon or spell, or you can use it defensively.

Fixed some random typos.

Step of the Wind is now a base class feature, and does not require Ki to use.